Remove AL2022 user-data ssh access workaround
This was added to workaround a bug in the packer amazon plugin:
https://github.com/hashicorp/packer-plugin-amazon/issues/172

The bug was fixed in plugin version 1.0.8, and we have now internally
migrated to plugin 1.1.5, so we no longer need the workaround.

External users can run `packer init -upgrade` to upgrade their packer
plugin versions if they still see this bug locally.