Add ecs-log-collector to AMI builds (excludes AL1) (#515)
ECS Log Collector is a script for gathering logs to troubleshooting Amazon ECS customer support cases. In some cases it is difficult to download the script into customer EC2 instance(s).

This change aims to include the script as part of the AMI build process to ensure that we remove one obstacle from customer's gathering logs for troubleshooting.

The version file, ECS_LOGS_COLLECTOR_VERSION, will be a version tracking file similar to NVIDA_DRIVER_VERSION with a git commit hash reflecting the ecs-logs-collector.sh version at time of AMI build. The script and version file can be observed in the /opt/amazon/ecs folder:
- ecs-logs-collector.sh - ECS Logs Collector shell script
- ECS_LOGS_COLLECTOR_VERSION - Contains the commit hash related to the latest changes for ECS Logs Collector shell script